+++ src/main/java/tool/updater/UpdaterCellRenderer.java
@@ -2,11 +2,9 @@ package tool.updater;
 
 import core.gui.theme.HOColorName;
 import core.gui.theme.ThemeManager;
-import java.awt.Component;
-import javax.swing.JButton;
-import javax.swing.JCheckBox;
-import javax.swing.JLabel;
-import javax.swing.JTable;
+
+import java.awt.*;
+import javax.swing.*;
 import javax.swing.table.TableCellRenderer;
 
 /**
@@ -20,29 +18,27 @@ public final class UpdaterCellRenderer implements TableCellRenderer {
 
 		double site = 0;
 		double ho = 0;
-		
-		
+		if (value == null) {
+			return new JLabel();
+		}
+
 		if (value instanceof JCheckBox) {
 			return (JCheckBox) value;
 		}
-		
+
 		try {
 			site = Double.parseDouble(((JLabel) table.getModel().getValueAt(row, 3)).getText());
-		} catch (Exception e) {
+		} catch (Exception ignored) {
 		}
 		try {
 			ho = Double.parseDouble(((JLabel) table.getModel().getValueAt(row, 2)).getText());
-		} catch (Exception e1) {
+		} catch (Exception ignored) {
 		}
 
-		boolean color = false;
-		if ((ho > 0) && (ho < site)) {
-			color = true;
-		}
+		boolean color = (ho > 0) && (ho < site);
 
-		if (value instanceof JButton) {
-			JButton b = (JButton) value;
-			b.setBorderPainted(false);
+        if (value instanceof JButton b) {
+            b.setBorderPainted(false);
 			if (color) {
 				b.setForeground(ThemeManager.getColor(HOColorName.LABEL_ERROR_FG));
 			}
@@ -57,6 +53,9 @@ public final class UpdaterCellRenderer implements TableCellRenderer {
 		JLabel label;
 		if (value instanceof JLabel) {
 			label = (JLabel) value;
+		} else if (value instanceof Color) {
+			label = new JLabel();
+			label.setBackground((Color) value);
 		} else {
 			label = new JLabel(value.toString());
 		}